Revert "chore: separate protos definition (#92)"
This reverts commit 29f1249c8c1c06d0f4bea3f1cd8bd277698b52ad.
