+++ .github/workflows/backend.yml
@@ -5,13 +5,11 @@ on:
     branches: [master]
     paths:
       - backend/**
-      - protos/**
       - .github/workflows/backend.yml
   pull_request:
     branches: [master]
     paths:
       - backend/**
-      - protos/**
       - .github/workflows/backend.yml
 
 jobs:
