+++ frontend/src/services/fund-cn/fund-cn.proto.ts
@@ -3,4 +3,4 @@ export {
   FundTypeCN,
   FundBasicInfoCN,
   FundValueWithInfoCN,
-} from '../../../../protos/fund-cn.proto';
+} from '../../../../backend/src/protos/fund-cn.proto';
